updated username in nav, moved function for featform in feed, renamed featreq
