+++ src/pages/Feed.js
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom'
 import { useEffect, useState } from 'react'
-import { GetFeats } from '../services/FeatAuth'
+import { GetFeats } from '../services/FeatReq'
 import FeatCard from '../components/FeatCard'
 import FeatForm from '../components/FeatForm'
 
@@ -27,6 +27,49 @@ const Feed = ({ feats, setFeats }) => {
     !active ? setActive(true) : setActive(false)
   }
 
+  const [featFormValues, setFeatFormValues] = useState({
+    type: '',
+    bodyPart: '',
+    intensity: 0,
+    description: '',
+    image: ''
+  })
+  const [emoji, setEmoji] = useState('')
+
+  const updateFeatFormValues = (e) => {
+    e.target.id === 'intensity'
+      ? setFeatFormValues({
+          ...featFormValues,
+          [e.target.id]: Number(e.target.value)
+        })
+      : setFeatFormValues({ ...featFormValues, [e.target.id]: e.target.value })
+    switch (featFormValues.intensity) {
+      case 0:
+        setEmoji('a')
+        break
+      case 1:
+        setEmoji('b')
+        break
+      case 2:
+        setEmoji('c')
+        break
+      case 3:
+        setEmoji('d')
+        break
+      case 4:
+        setEmoji('e')
+        break
+      case 5:
+        setEmoji('f')
+        break
+      default:
+    }
+  }
+
+  const submitFeatForm = async (e) => {
+    e.preventDefault()
+  }
+
   return (
     <div>
       <div className="feat-grid">
@@ -40,7 +83,13 @@ const Feed = ({ feats, setFeats }) => {
         Create Feat
       </button>
       <div style={{ display: `${formDisplay}` }}>
-        <FeatForm displayCreateFeat={displayCreateFeat} />
+        <FeatForm
+          displayCreateFeat={displayCreateFeat}
+          featFormValues={featFormValues}
+          updateFeatFormValues={updateFeatFormValues}
+          submitFeatForm={submitFeatForm}
+          emoji={emoji}
+        />
       </div>
     </div>
   )
